removed redundant file
